Allow an alternate startup file independent of bootloader builds
In support of bootloader builds the nuttx_v3 builds previously had a
facility to set the start up files: Non Bootloader STM32 builds uses
the common vectors facility in NuttX. The bootloaders use a fixed
minimum set of vectors.

Since other architectures my need to include a start up file or
set of files, this PR allow an alternate startup file to be
selected independent of if this is bootloader build.